Fix a bug for BGP Open Message
While tried with FRR with faucet for BGP, it runns into problem
while parsing BGP capability in OPEN msg, the while loop cannot
break correctly.

this happens when the capability_code is not in capability_keys,
and the capability_length is not 0, the second while loop will read
data from stream, but it should not.

Signed-off-by: Yong Xu <yong.xu@corigine.com>